update score type to be float to match dmd format

+++ src/search_disk.h
@@ -95,7 +95,7 @@ QueryIterator* SearchDisk_NewWildcardIterator(RedisSearchDiskIndexSpec *index, d
  * @param maxFreq Maximum term frequency in the document
  * @return New document ID, or 0 on error/duplicate
  */
-t_docId SearchDisk_PutDocument(RedisSearchDiskIndexSpec *handle, const char *key, double score, uint32_t flags, uint32_t maxFreq);
+t_docId SearchDisk_PutDocument(RedisSearchDiskIndexSpec *handle, const char *key, float score, uint32_t flags, uint32_t maxFreq);
 
 /**
  * @brief Get document metadata by document ID
